+++ src/view/offer.js
@@ -1,0 +1,36 @@
+import {createElement} from "../helpers/create-element";
+
+const createOfferTemplate = (offer) => {
+  const {name, price} = offer;
+
+  return `
+    <li class="event__offer">
+      <span class="event__offer-title">${name}</span>
+      &plus;&euro;&nbsp;
+      <span class="event__offer-price">${price}</span>
+    </li>
+  `;
+};
+
+export default class Offer {
+  constructor(offer) {
+    this._offer = offer;
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createOfferTemplate(this._offer);
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
